move footer into own component
